OAK-1W/requirements.txt
opencv-python
depthai
pandas
google-genai
openpyxl